refactor: unify bridge + nuxt 3 implementations
