+++ src/runtime/composables.ts
@@ -1,5 +1,5 @@
 import { useState } from '#app'
-import type { ColorModeInstance } from '../types'
+import type { ColorModeInstance } from './types'
 
 export const useColorMode = () => {
   return useState('color-mode').value as ColorModeInstance
