--- src/runtime/vue3/plugin.server.ts
@@ -1,50 +1,0 @@
-import { preference, script } from '#color-mode-options'
-import { defineNuxtPlugin } from '#app'
-import { addRouteMiddleware, useMeta, useState } from '#imports'
-import { reactive } from 'vue'
-
-import type { ColorModeInstance } from '../types'
-
-export default defineNuxtPlugin((nuxtApp) => {
-  // Workaround until we have support in vueuse/head
-  if ('renderMeta' in nuxtApp.ssrContext) {
-    const originalRender = nuxtApp.ssrContext.renderMeta
-    nuxtApp.ssrContext.renderMeta = async () => {
-      const result = await originalRender()
-      result.bodyScripts = `<script>${script}</script>` + (result.bodyScripts || '')
-      return result
-    }
-  }
-
-  const colorMode = useState<ColorModeInstance>('color-mode', () => reactive({
-    preference,
-    value: preference,
-    unknown: true,
-    forced: false
-  })).value
-
-  addRouteMiddleware('color-mode', (to, from) => {
-    const forcedColorMode = to.meta.colorMode
-
-    if (forcedColorMode && forcedColorMode !== 'system') {
-      colorMode.value = forcedColorMode
-      colorMode.forced = true
-
-      useMeta({
-        bodyAttrs: {
-          'data-color-mode-forced': forcedColorMode
-        }
-      })
-    } else {
-      if (forcedColorMode === 'system') {
-        console.warn('You cannot force the colorMode to system at the page level.')
-      }
-    }
-  }, { global: true })
-
-  return {
-    provide: {
-      colorMode
-    }
-  }
-})
